# cmd/app-core/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample service server.",
        "title": "Swagger Example API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/service/api/v1",
    "paths": {
        "/cards": {
            "get": {
                "description": "Get info API",
                "tags": [
                    "card"
                ],
                "summary": "Get info API",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "post Card",
                "tags": [
                    "card"
                ],
                "summary": "post Card API",
                "parameters": [
                    {
                        "description": "CardUpdate",
                        "name": "CardUpdate",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CardUpdate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/cards/{id}": {
            "put": {
                "description": "put Card API",
                "tags": [
                    "card"
                ],
                "summary": "put Card API",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Card ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "CardUpdate",
                        "name": "CardUpdate",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CardUpdate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete Card API",
                "tags": [
                    "card"
                ],
                "summary": "Delete Card API",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Card ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/info": {
            "get": {
                "description": "Get info API",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get info API",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "patch": {
                "description": "just test only , not really update",
                "summary": "test patch InfoUpdate API",
                "parameters": [
                    {
                        "description": "InfoUpdate",
                        "name": "InfoUpdate",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.InfoUpdate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.CardUpdate": {
            "type": "object",
            "required": [
                "title"
            ],
            "properties": {
                "content": {
                    "type": "string",
                    "example": "finished CRUD APIs"
                },
                "title": {
                    "type": "string",
                    "example": "learn golang"
                }
            }
        },
        "models.InfoUpdate": {
            "type": "object",
            "required": [
                "description",
                "name"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        }
    }
}